+++ packages/reactabular/CHANGELOG.md
@@ -9,6 +9,7 @@
 
   * Bug fix - Allow `tree.toggleChildren` to work without column `props` defined.
   * Feature - Add `tree.getImmediateChildren`.
+  * Feature - Add `tree.moveRows`.
 
 ## reactabular-utils